CORE-15909: moving logic of fetching group params to utxo ledger tx factory
