+++ components/ledger/ledger-utxo-flow/src/main/kotlin/net/corda/ledger/utxo/flow/impl/transaction/verifier/UtxoLedgerTransactionVerificationServiceImpl.kt
@@ -1,15 +1,12 @@
 package net.corda.ledger.utxo.flow.impl.transaction.verifier
 
 import io.micrometer.core.instrument.Timer
-import net.corda.crypto.core.parseSecureHash
 import net.corda.flow.external.events.executor.ExternalEventExecutor
 import net.corda.flow.fiber.metrics.recordSuspendable
 import net.corda.ledger.common.data.transaction.TransactionMetadataInternal
 import net.corda.ledger.utxo.data.transaction.TransactionVerificationStatus
 import net.corda.ledger.utxo.data.transaction.UtxoLedgerTransactionContainer
 import net.corda.ledger.utxo.data.transaction.UtxoLedgerTransactionInternal
-import net.corda.ledger.utxo.flow.impl.groupparameters.GroupParametersServiceInternal
-import net.corda.ledger.utxo.flow.impl.persistence.UtxoLedgerGroupParametersPersistenceService
 import net.corda.ledger.utxo.flow.impl.transaction.verifier.external.events.TransactionVerificationExternalEventFactory
 import net.corda.ledger.utxo.flow.impl.transaction.verifier.external.events.TransactionVerificationParameters
 import net.corda.ledger.utxo.flow.impl.groupparameters.verifier.SignedGroupParametersVerifier
@@ -39,10 +36,6 @@ class UtxoLedgerTransactionVerificationServiceImpl @Activate constructor(
     private val externalEventExecutor: ExternalEventExecutor,
     @Reference(service = SerializationService::class)
     private val serializationService: SerializationService,
-    @Reference(service = UtxoLedgerGroupParametersPersistenceService::class)
-    private val utxoLedgerGroupParametersPersistenceService: UtxoLedgerGroupParametersPersistenceService,
-    @Reference(service = GroupParametersServiceInternal::class)
-    private val groupParametersService: GroupParametersServiceInternal,
     @Reference(service = CurrentSandboxGroupContext::class)
     private val currentSandboxGroupContext: CurrentSandboxGroupContext,
     @Reference(service = SignedGroupParametersVerifier::class)
@@ -53,7 +46,7 @@ class UtxoLedgerTransactionVerificationServiceImpl @Activate constructor(
     override fun verify(transaction: UtxoLedgerTransaction) {
         recordSuspendable(::transactionVerificationFlowTimer) @Suspendable {
 
-            val signedGroupParameters = fetchAndVerifySignedGroupParameters(transaction)
+            val signedGroupParameters = verifySignedGroupParameters(transaction)
             verifyNotaryAllowed(transaction, signedGroupParameters)
 
             val verificationResult = externalEventExecutor.execute(
@@ -76,23 +69,12 @@ class UtxoLedgerTransactionVerificationServiceImpl @Activate constructor(
     }
 
     @Suspendable
-    private fun fetchAndVerifySignedGroupParameters(transaction: UtxoLedgerTransaction): SignedGroupParameters {
-        val membershipGroupParametersHashString = transaction.getMembershipGroupParametersHash()
-
-        val currentGroupParameters = groupParametersService.currentGroupParameters
+    private fun verifySignedGroupParameters(transaction: UtxoLedgerTransaction): SignedGroupParameters {
         val signedGroupParameters =
-            if (currentGroupParameters.hash.toString() == membershipGroupParametersHashString) {
-                currentGroupParameters
-            } else {
-                val membershipGroupParametersHash = parseSecureHash(membershipGroupParametersHashString)
-                utxoLedgerGroupParametersPersistenceService.find(membershipGroupParametersHash)
-            }
-        requireNotNull(signedGroupParameters) {
-            "Signed group parameters $membershipGroupParametersHashString related to the transaction ${transaction.id} cannot be accessed."
-        }
+            transaction.groupParameters
         signedGroupParametersVerifier.verify(
             transaction,
-            signedGroupParameters
+            signedGroupParameters as SignedGroupParameters
         )
         return signedGroupParameters
     }
@@ -113,12 +95,6 @@ class UtxoLedgerTransactionVerificationServiceImpl @Activate constructor(
         }
 
     private fun serialize(payload: Any) = ByteBuffer.wrap(serializationService.serialize(payload).bytes)
-
-    private fun UtxoLedgerTransaction.getMembershipGroupParametersHash(): String {
-        return requireNotNull((metadata as TransactionMetadataInternal).getMembershipGroupParametersHash()) {
-            "Membership group parameters hash cannot be found in the transaction metadata."
-        }
-    }
 
     private fun transactionVerificationFlowTimer(): Timer {
         return CordaMetrics.Metric.Ledger.TransactionVerificationFlowTime
